refactor: remove `controller/panel.ts`
